+++ src/axom/inlet/tests/inlet_test_utils.hpp
@@ -12,6 +12,7 @@
 #include "gtest/gtest.h"
 
 #include "axom/inlet/YAMLReader.hpp"
+#include "axom/inlet/JSONReader.hpp"
 
 #ifdef AXOM_USE_SOL
   #include "axom/inlet/LuaReader.hpp"
@@ -25,14 +26,14 @@ namespace detail
 {
 /*!
  *******************************************************************************
- * \class LuaToYAML
- * \brief A converter class that translates Lua to YAML
+ * \class LuaTranslator
+ * \brief A converter class that translates Lua to YAML or JSON
  * \note This class is not a fully functional Lua parser and should not be treated
  * as such.  It is designed only to parse the subset of Lua that maps to valid
- * YAML.
+ * YAML/JSON.
  *******************************************************************************
  */
-class LuaToYAML
+class LuaTranslator
 {
 public:
   /*!
@@ -43,7 +44,17 @@ public:
    * responsibility of the caller to pass a valid Lua string.
    *****************************************************************************
    */
-  static std::string convert(const std::string& luaString);
+  static std::string convertYAML(const std::string& luaString);
+
+  /*!
+   *****************************************************************************
+   * \brief Converts a Lua string to JSON
+   * \param [in] luaString The string to convert
+   * \note This function does not check for syntactic validity.  It is the
+   * responsibility of the caller to pass a valid Lua string.
+   *****************************************************************************
+   */
+  static std::string convertJSON(const std::string& luaString);
 
 private:
   /*!
@@ -83,14 +94,21 @@ inline std::string fromLuaTo(const std::string& luaString)
 template <>
 inline std::string fromLuaTo<axom::inlet::YAMLReader>(const std::string& luaString)
 {
-  return LuaToYAML::convert(luaString);
+  return LuaTranslator::convertYAML(luaString);
+}
+/// \overload
+template <>
+inline std::string fromLuaTo<axom::inlet::JSONReader>(const std::string& luaString)
+{
+  return LuaTranslator::convertJSON(luaString);
 }
 
 #ifdef AXOM_USE_SOL
 using ReaderTypes =
-  ::testing::Types<axom::inlet::LuaReader, axom::inlet::YAMLReader>;
+  ::testing::Types<axom::inlet::LuaReader, axom::inlet::YAMLReader, axom::inlet::JSONReader>;
 #else
-using ReaderTypes = ::testing::Types<axom::inlet::YAMLReader>;
+using ReaderTypes =
+  ::testing::Types<axom::inlet::YAMLReader, axom::inlet::JSONReader>;
 #endif
 
 }  // namespace detail